added parameter to recert.sh

+++ recert.sh
@@ -2,8 +2,13 @@
 
 # https://www.wissel.net/blog/2018/03/letsencrypt-java-keystore.html
 
+if [ "$1" = "" ]; then
+  echo "Need to pass in argument 'domain', e.g. schedge.a1liu.com"
+  exit
+f
+
 sudo certbot renew
-cat /etc/letsencrypt/live/schedge.torchnyu.com/*.pem > fullcert.pem
+cat "/etc/letsencrypt/live/$1/*.pem" > fullcert.pem
 openssl pkcs12 -export -out fullcert.pkcs12 -in fullcert.pem
 rm fullcert.pem
 cp src/main/resources/empty.jks keystore.jks
